backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo[srv]==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
httpx==0.25.2
python-multipart==0.0.6
email-validator
pydantic-settings==2.1.0
python-dotenv==1.0.0
